makestream -> corostream
makestream used a standard futures channel to create a
stream/sink pair, and ran the content-generating task
seperately, writing to the sink. that task usually ends up
on a different thread, so that results in a lot of inter-thread
traffic.

Replace it with 'corostream', that runs it all in the
same task.