pass &Request to internal handlers
so that the Request remains available in handle2() after the
handler has been called. on error we can then pass the Request
back _in_ the error struct.

that does mean we'll have to define our own errortype instead
of io::Result. we can then return that directly, or we can implement
some trait for io::Result that downcasts the internal error to our
own errortype.