localfs: make it work with task::spawn_blocking()
Up until now it only worked with task::block_in_place, which makes
borrowing issues a lot easier. that however does not work with the
basic scheduler, and thus with actix-web.

That should all work now. The only issue left is that we copy
the read/write buffers at the moment, we should just move ownership.
Perhaps a special buffer type that wraps Option<Vec<u8>>.